feat(operators): create a separate page for operator management and move the operator management code that was on the dashboard to a separate page, maintaining all existing functionality.
